Couple of method to identify playlist and drama serial. Drama serial will not refresh on season change.

+++ app/src/main/java/com/banglalink/toffee/ui/player/PlaylistManager.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import com.banglalink.toffee.model.ChannelInfo
 
 class PlaylistManager {
-    var playlistId: Int = -1
+    var playlistId: Long = -1L
     private val playList: MutableList<ChannelInfo> = mutableListOf()
     private var playlistIndex = -1
